requirements.txt
kivy==2.1.0
kivymd==1.0.2
pyserial==3.5
